SRC Personnel Overcounted or Missing
Last commit had an oversight.  We don't sum the personnel count
across all of the supply records.  That doesn't make sense.  Also,
there are some SRCs without an AC record, so we can't filter on
that.  Assume that the personnel count for all records for one
SRC are equal.  Then we can just take the first personnel count
that we find.