Automated creation of COMFORTER product for sponsor
Given the original comforter output table from comforter.core,
made a new namespace comforter.summary in which we can generate
an Excel workbook that contains the original output table in
one sheet and a summary table of the total additional cost
and additional personnel in another sheet.  This workbook is the
product we sent directly to the sponsor for 22-26.  A blurb
exists on the high side to accompany the summary table.